# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Siuji TOEFL exam application.",
        "title": "Siuji API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/periods/{period_public_id}/participants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "List participants in a period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name/email/nim",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponsePaginated"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ParticipantResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates the user account if the email doesn't exist yet. Password is auto-generated from NIM.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Add participant to period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Participant data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddParticipantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ParticipantResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/periods/{period_public_id}/participants/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Columns (in order, header row skipped): Name, Email, NIM, University.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Import participants from Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Excel file (.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ImportParticipantResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/periods/{period_public_id}/participants/{user_public_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Get participant detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User public ID",
                        "name": "user_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ParticipantResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partial update — only non-empty fields are applied. Changing NIM regenerates the password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Update participant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User public ID",
                        "name": "user_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateParticipantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ParticipantResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Remove participant from period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User public ID",
                        "name": "user_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AddParticipantRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "nim"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                }
            }
        },
...
        "model.ImportErrorDetail": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        },
        "model.ImportParticipantResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ImportErrorDetail"
                    }
                },
                "total_imported": {
                    "type": "integer"
                },
                "total_skipped": {
                    "type": "integer"
                }
            }
        },
...
        "model.ParticipantResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "period_public_id": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                }
            }
        },
...
        "model.UpdateParticipantRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "registered",
                        "started",
                        "completed"
                    ]
                },
                "university": {
                    "type": "string"
                }
            }
        },
...
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.PaginationMeta": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string"
                },
                "total_datas": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ResponseNoData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ResponsePaginated": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.PaginationMeta"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TempToken": {
            "description": "Type \"Bearer\" followed by a space and temp JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}